+++ app/src/main/java/com/example/airlinesapp/di/network/ApiService.kt
@@ -7,7 +7,6 @@ import io.reactivex.Single
 import retrofit2.http.GET
 import retrofit2.http.Query
 
-
 interface ApiService {
 
     @GET("airlines")
