+++ app/src/main/java/com/example/airlinesapp/util/Constants.kt
@@ -3,10 +3,8 @@ package com.example.airlinesapp.util
 object Constants {
 
     const val SPLASHING_TIME = 2000
-    const val BASE_URL = "https://api.instantwebtools.net/v1/"
     const val PASSENGERS_PER_PAGE = 10
     const val FIRST_PAGE = 1
-
     const val LOADING = "Loading"
     const val LOADED = "Loaded"
     const val CHECK_NETWORK_ERROR = "Check your network connection."
