+++ app/src/main/res/layout/fragment_airlines.xml
@@ -13,7 +13,8 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             tools:layout_editor_absoluteX="0dp"
-            tools:layout_editor_absoluteY="0dp" />
+            tools:layout_editor_absoluteY="0dp"
+            tools:listitem="@layout/single_airline_view"/>
 
         <ProgressBar
             android:id="@+id/progressBar"
